No longer always removing and readding all views on every change.
It turns out that removing and readding a view to the tree is slow.
During the time that the views are missing from the tree, keyboard events don't get sent to what is supposed to be the active textbox.
By keeping the views in the tree in cases when views are merely resizing, text inputs don't get skipped.