Making the ScrollLayout consistent
Previously it was possible for the ScrollLayout to violate the assumptions required by a LayoutCache,
where a layout query might tighten its bounds and receive a more desirable response.

This was because the ScrollLayout was checking a set of layouts in multiples of the query height.
Now the ScrollLayout is the union of a ContainerLayout and an ImageLayout, so it's consistent.
We still want to make it smarter about not asking for more space than it really needs but that's less necessary.